Enum supply_chain_io::SupplyChainAction

pub enum SupplyChainAction {

    Produce {
        name: String,
        description: String,
    },
    PutUpForSaleByProducer {
        item_id: ItemId,
        price: u128,
    },
    PurchaseByDistributor {
        item_id: ItemId,
        delivery_time: u64,
    },
    ApproveByProducer {
        item_id: ItemId,
        approve: bool,
    },
    ShipByProducer(ItemId),
    ReceiveByDistributor(ItemId),
    ProcessByDistributor(ItemId),
    PackageByDistributor(ItemId),
    PutUpForSaleByDistributor {
        item_id: ItemId,
        price: u128,
    },
    PurchaseByRetailer {
        item_id: ItemId,
        delivery_time: u64,
    },
    ApproveByDistributor {
        item_id: ItemId,
        approve: bool,
    },
    ShipByDistributor(ItemId),
    ReceiveByRetailer(ItemId),
    PutUpForSaleByRetailer {
        item_id: ItemId,
        price: u128,
    },
    PurchaseByConsumer(ItemId),

}

Variants

Produce

Fields

name: String

An item’s name.

description: String

An item’s description.

Produces one item with a name and description and replies with its ID.

Transfers created NFT for an item to a producer.

Requirements

msg::source() must be a producer in a supply chain.

On success, returns SupplyChainEvent::Produced.

PutUpForSaleByProducer

Fields

item_id: ItemId

An item’s ID.

price: u128

An item’s price.

Puts an item up for a sale to a distributor for a given price on behalf of a producer.

Transfers item’s NFT to a supply chain.

Requirements

msg::source() must be a producer in a supply chain and a producer of this item.

Item’s ItemState must be Produced.

On success, returns SupplyChainEvent::ForSaleByProducer.

PurchaseByDistributor

Fields

item_id: ItemId

An item’s ID.

delivery_time: u64

A time in milliseconds for which a producer must deliver an item. A countdown starts after SupplyChainAction::ShipByProducer is executed.

Purchases an item from a producer on behalf of a distributor.

Transfers tokens for purchasing an item to a supply chain until an item is received (by SupplyChainAction::ReceiveByDistributor).

Note that an item’s producer must approve or not this purchase by SupplyChainAction::ApproveByProducer.

Requirements

msg::source() must be a distributor in a supply chain.

Item’s ItemState must be ForSaleByProducer.

On success, returns SupplyChainEvent::PurchasedByDistributor.

ApproveByProducer

Fields

item_id: ItemId

An item’s ID.

approve: bool

Yes (true) or no (false).

Approves or not a purchase from a distributor on behalf of a producer.

If a purchase is approved, then item’s ItemState changes to ApprovedByProducer and an item can be shipped.

If a purchase is not approved, then tokens for a purchase are refunded to an item’s producer.

Requirements

msg::source() must be a producer in a supply chain and a producer of this item.

Item’s ItemState must be PurchasedByDistributor.

On success, returns SupplyChainEvent::ApprovedByProducer.

ShipByProducer(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Starts shipping a purchased item to a distributor on behalf of a producer.

Starts a countdown for a delivery time that was specified for this item in SupplyChainAction::PurchaseByDistributor.

Requirements

msg::source() must be a producer in a supply chain and a producer of this item.

Item’s ItemState must be PurchasedByDistributor.

On success, returns SupplyChainEvent::ShippedByProducer.

ReceiveByDistributor(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Receives a shipped item from a producer on behalf of a distributor.

Depending on a counted delivery time, transfers tokens for purchasing an item from a supply chain to a producer or as a penalty for being late refunds some or all of them to a distributor.

Transfers item’s NFT to a distributor.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be ShippedByProducer.

On success, returns SupplyChainEvent::ReceivedByDistributor.

ProcessByDistributor(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Processes a received item from a producer on behalf of a distributor.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be ReceivedByDistributor.

On success, returns SupplyChainEvent::ProcessedByDistributor.

PackageByDistributor(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Packages a processed item on behalf of a distributor.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be ProcessedByDistributor.

On success, returns SupplyChainEvent::PackagedByDistributor.

PutUpForSaleByDistributor

Fields

item_id: ItemId

An item’s ID.

price: u128

An item’s price.

Puts a packaged item up for a sale to a retailer for a given price on behalf of a distributor.

Transfers item’s NFT to a supply chain.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be PackagedByDistributor.

On success, returns SupplyChainEvent::ForSaleByDistributor.

PurchaseByRetailer

Fields

item_id: ItemId

An item’s ID.

delivery_time: u64

A time in milliseconds for which a distributor must deliver an item. A countdown starts after SupplyChainAction::ShipByDistributor is executed.

Purchases an item from a distributor on behalf of a retailer.

Transfers tokens for purchasing an item to a supply chain until an item is received (by SupplyChainAction::ReceiveByRetailer).

Requirements

msg::source() must be a retailer in a supply chain.

Item’s ItemState must be ForSaleByDistributor.

On success, returns SupplyChainEvent::PurchasedByRetailer.

ApproveByDistributor

Fields

item_id: ItemId

An item’s ID.

approve: bool

Yes (true) or no (false).

Approves or not a purchase from a retailer on behalf of a distributor.

If a purchase is approved, then item’s ItemState changes to ApprovedByDistributor and an item can be shipped.

If a purchase is not approved, then tokens for a purchase are refunded to an item’s distributor.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be PurchasedByRetailer.

On success, returns SupplyChainEvent::ApprovedByDistributor.

ShipByDistributor(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Starts shipping a purchased item to a retailer on behalf of a distributor.

Starts a countdown for a delivery time that was specified for this item in SupplyChainAction::PurchaseByRetailer.

Requirements

msg::source() must be a distributor in a supply chain and a distributor of this item.

Item’s ItemState must be PurchasedByRetailer.

On success, returns SupplyChainEvent::ShippedByDistributor.

ReceiveByRetailer(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Receives a shipped item from a distributor on behalf of a retailer.

Depending on a counted delivery time, transfers tokens for purchasing an item from a supply chain to a distributor or as a penalty for being late refunds some or all of them to a retailer.

Transfers item’s NFT to a retailer.

Requirements

msg::source() must be a retailer in a supply chain and a retailer of this item.

Item’s ItemState must be ShippedByDistributor.

On success, returns SupplyChainEvent::ReceivedByRetailer.

PutUpForSaleByRetailer

Fields

item_id: ItemId

An item’s ID.

price: u128

An item’s price.

Puts a received item from a distributor up for a sale to a consumer for a given price on behalf of a retailer.

Transfers item’s NFT to a supply chain.

Requirements

msg::source() must be a retailer in a supply chain and a retailer of this item.

Item’s ItemState must be ReceivedByRetailer.

On success, returns SupplyChainEvent::ForSaleByRetailer.

PurchaseByConsumer(ItemId)

Tuple Fields

0: ItemId

An item’s ID.

Purchases an item from a retailer.

Transfers tokens for purchasing an item to its retailer.

Transfers item’s NFT to a consumer.

Requirements

Item’s ItemState must be ForSaleByRetailer.

On success, returns SupplyChainEvent::PurchasedByConsumer.