Enum supply_chain_io::SupplyChainEvent

pub enum SupplyChainEvent {

    Produced(ItemId),
    ForSaleByProducer(ItemId),
    PurchasedByDistributor(ItemId),
    ApprovedByProducer(ItemId),
    ShippedByProducer(ItemId),
    ReceivedByDistributor(ItemId),
    ProcessedByDistributor(ItemId),
    PackagedByDistributor(ItemId),
    ForSaleByDistributor(ItemId),
    PurchasedByRetailer(ItemId),
    ApprovedByDistributor(ItemId),
    ShippedByDistributor(ItemId),
    ReceivedByRetailer(ItemId),
    ForSaleByRetailer(ItemId),
    PurchasedByConsumer(ItemId),

}

Variants

Produced(ItemId)

Tuple Fields

0: ItemId

An ID of a produced item.

ForSaleByProducer(ItemId)

Tuple Fields

0: ItemId

An ID of an item put up for a sale by a producer.

PurchasedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item purchased by a distributor.

ApprovedByProducer(ItemId)

Tuple Fields

0: ItemId

An ID of an item approved by a producer for a purchase.

ShippedByProducer(ItemId)

Tuple Fields

0: ItemId

An ID of an item shipped by a producer.

ReceivedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item received by a distributor.

ProcessedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item processed by a distributor.

PackagedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item packaged by a distributor.

ForSaleByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item put up for a sale by a distributor.

PurchasedByRetailer(ItemId)

Tuple Fields

0: ItemId

An ID of an item purchased by a retailer.

ApprovedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item approved by a retailer for a purchase.

ShippedByDistributor(ItemId)

Tuple Fields

0: ItemId

An ID of an item shipped by a distributor.

ReceivedByRetailer(ItemId)

Tuple Fields

0: ItemId

An ID of an item received by a retailer.

ForSaleByRetailer(ItemId)

Tuple Fields

0: ItemId

An ID of an item put up for a sale by a retailer.

PurchasedByConsumer(ItemId)

Tuple Fields

0: ItemId

An ID of an item purchased by a consumer.